+++ ui/src/dashboards/components/dashboard_index/DashboardCard.tsx
@@ -6,15 +6,14 @@ import {withRouter, WithRouterProps} from 'react-router'
 // Components
 import {IconFont, ComponentColor} from '@influxdata/clockface'
 import {ResourceList, Context} from 'src/clockface'
-import FeatureFlag from 'src/shared/components/FeatureFlag'
 import InlineLabels from 'src/shared/components/inlineLabels/InlineLabels'
 
 // Actions
-import {createLabelAJAX} from 'src/labels/api'
 import {
   addDashboardLabelsAsync,
   removeDashboardLabelsAsync,
 } from 'src/dashboards/actions/v2'
+import {createLabel as createLabelAsync} from 'src/labels/actions'
 
 // Types
 import {Organization} from 'src/types/v2'
@@ -41,6 +40,7 @@ interface StateProps {
 interface DispatchProps {
   onAddDashboardLabels: typeof addDashboardLabelsAsync
   onRemoveDashboardLabels: typeof removeDashboardLabelsAsync
+  onCreateLabel: typeof createLabelAsync
 }
 
 type Props = PassedProps & DispatchProps & StateProps & WithRouterProps
@@ -98,15 +98,13 @@ class DashboardCard extends PureComponent<Props> {
 
     return (
       <Context>
-        <FeatureFlag>
-          <Context.Menu icon={IconFont.CogThick}>
-            <Context.Item
-              label="Export"
-              action={this.handleExport}
-              value={dashboard}
-            />
-          </Context.Menu>
-        </FeatureFlag>
+        <Context.Menu icon={IconFont.CogThick}>
+          <Context.Item
+            label="Export"
+            action={this.handleExport}
+            value={dashboard}
+          />
+        </Context.Menu>
         <Context.Menu
           icon={IconFont.Duplicate}
           color={ComponentColor.Secondary}
@@ -160,13 +158,13 @@ class DashboardCard extends PureComponent<Props> {
     onRemoveDashboardLabels(dashboard.id, [label])
   }
 
-  private handleCreateLabel = async (label: ILabel): Promise<ILabel> => {
+  private handleCreateLabel = async (label: ILabel): Promise<void> => {
     try {
-      const newLabel = await createLabelAJAX(label)
+      await this.props.onCreateLabel(label.name, label.properties)
+
       // notify success
-      return newLabel
     } catch (err) {
-      console.log(err)
+      console.error(err)
       // notify of fail
       throw err
     }
@@ -190,6 +188,7 @@ const mstp = ({labels}: AppState): StateProps => {
 }
 
 const mdtp: DispatchProps = {
+  onCreateLabel: createLabelAsync,
   onAddDashboardLabels: addDashboardLabelsAsync,
   onRemoveDashboardLabels: removeDashboardLabelsAsync,
 }