+++ ui/src/tasks/components/TaskCard.tsx
@@ -6,14 +6,11 @@ import {withRouter, WithRouterProps} from 'react-router'
 // Components
 import {SlideToggle, ComponentSize} from '@influxdata/clockface'
 import {ResourceList, Context} from 'src/clockface'
-import FeatureFlag from 'src/shared/components/FeatureFlag'
 import InlineLabels from 'src/shared/components/inlineLabels/InlineLabels'
-
-// API
-import {createLabelAJAX} from 'src/labels/api'
 
 // Actions
 import {addTaskLabelsAsync, removeTaskLabelsAsync} from 'src/tasks/actions/v2'
+import {createLabel as createLabelAsync} from 'src/labels/actions'
 
 // Types
 import {ComponentColor} from '@influxdata/clockface'
@@ -42,6 +39,7 @@ interface StateProps {
 interface DispatchProps {
   onAddTaskLabels: typeof addTaskLabelsAsync
   onRemoveTaskLabels: typeof removeTaskLabelsAsync
+  onCreateLabel: typeof createLabelAsync
 }
 
 type Props = PassedProps & StateProps & DispatchProps
@@ -90,9 +88,7 @@ export class TaskCard extends PureComponent<Props & WithRouterProps> {
     return (
       <Context>
         <Context.Menu icon={IconFont.CogThick}>
-          <FeatureFlag>
-            <Context.Item label="Export" action={this.handleExport} />
-          </FeatureFlag>
+          <Context.Item label="Export" action={this.handleExport} />
           <Context.Item label="View Task Runs" action={this.handleViewRuns} />
           <Context.Item label="Run Task" action={onRunTask} value={task.id} />
         </Context.Menu>
@@ -170,11 +166,9 @@ export class TaskCard extends PureComponent<Props & WithRouterProps> {
     onRemoveTaskLabels(task.id, [label])
   }
 
-  private handleCreateLabel = async (label: ILabel): Promise<ILabel> => {
+  private handleCreateLabel = async (label: ILabel): Promise<void> => {
     try {
-      const newLabel = await createLabelAJAX(label)
-
-      return newLabel
+      await this.props.onCreateLabel(label.name, label.properties)
     } catch (err) {
       throw err
     }
@@ -220,6 +214,7 @@ const mstp = ({labels}: AppState): StateProps => {
 }
 
 const mdtp: DispatchProps = {
+  onCreateLabel: createLabelAsync,
   onAddTaskLabels: addTaskLabelsAsync,
   onRemoveTaskLabels: removeTaskLabelsAsync,
 }